Update createChar()'s charmap paramter type to byte[]
The type of charmap parameter of createChar() changed in LiquidCrystal_PCF8574 1.3.0 to byte[].
Turn charmaps to byte[], instead of int[].
Fix LiquidCrystal_PCF8574 mock, too.

+++ water_system/LcdGlyphs.h
@@ -3,20 +3,20 @@
 
 #define LCD_LINES_PER_GLYPH 8
 
-extern const int plant[LCD_LINES_PER_GLYPH] PROGMEM;
-extern const int rain_plant[LCD_LINES_PER_GLYPH] PROGMEM;
-extern const int right_arrow[LCD_LINES_PER_GLYPH] PROGMEM;
-extern const int level_up[LCD_LINES_PER_GLYPH] PROGMEM;
-extern const int level_mid[LCD_LINES_PER_GLYPH] PROGMEM;
-extern const int level_low[LCD_LINES_PER_GLYPH] PROGMEM;
-extern const int burger_menu[LCD_LINES_PER_GLYPH] PROGMEM;
-extern const int skull[LCD_LINES_PER_GLYPH] PROGMEM;
+extern const byte plant[LCD_LINES_PER_GLYPH] PROGMEM;
+extern const byte rain_plant[LCD_LINES_PER_GLYPH] PROGMEM;
+extern const byte right_arrow[LCD_LINES_PER_GLYPH] PROGMEM;
+extern const byte level_up[LCD_LINES_PER_GLYPH] PROGMEM;
+extern const byte level_mid[LCD_LINES_PER_GLYPH] PROGMEM;
+extern const byte level_low[LCD_LINES_PER_GLYPH] PROGMEM;
+extern const byte burger_menu[LCD_LINES_PER_GLYPH] PROGMEM;
+extern const byte skull[LCD_LINES_PER_GLYPH] PROGMEM;
 
 class LCDGlyph {
     private:
         int _custom_glyph_location;
 
     public:
-        LCDGlyph(LiquidCrystal_PCF8574 &lcd, int location, const int charmap_P[] PROGMEM);
+        LCDGlyph(LiquidCrystal_PCF8574 &lcd, int location, const byte charmap_P[] PROGMEM);
         int location();
 };
